refactor: extract reducer space lifecycle
